requirements.txt
PySide6>=6.7.0,<7.0.0
SQLAlchemy>=2.0.30,<3.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
openpyxl>=3.1.0,<4.0.0
bcrypt<4.0
PySide6-Addons>=6.7.0,<7.0.0
pandas>=2.1.3,<3.0.0
python-dotenv>=1.0.1,<2.0.0